fix BMP render function

+++ lib/GfxUi/GfxUi.h
@@ -43,7 +43,7 @@ https://github.com/Bodmer/TFT_eSPI
 class GfxUi {
   public:
     GfxUi(TFT_eSPI * tft);
-    void drawBmp(String filename, uint8_t x, uint16_t y);
+    void drawBmp(String filename, uint16_t x, uint16_t y);
     void drawProgressBar(uint16_t x, uint16_t y, uint16_t w, uint16_t h, uint8_t percentage, uint16_t frameColor, uint16_t barColor);
     void jpegInfo();
     void drawJpeg(const char *filename, int xpos, int ypos);
